+++ frontend/src/utils/validateBookCreatePayload.ts
@@ -1,0 +1,9 @@
+import { PayloadType } from '../api/hooks/usePostBooks';
+
+export const validateBookCreatePayloads = (payload: PayloadType) => {
+	const values = Object.values(payload);
+	for (let i = 0; i < values.length; i++) {
+		if (!values[i]) return false;
+	}
+	return true;
+};
